+++ tests/testthat/test_install.R
@@ -48,7 +48,6 @@ test_that(
 )
 
 
-
 test_that(
   "download and install UNGAMINI",
   {