+++ NEWS.md
@@ -20,6 +20,7 @@ cwbtools 0.3.2
 * The function `cwb_get_bindir()` now introduces an argument `bindir`.
 * Argument `compress` of `p_attribute_encode(` now has default value `FALSE` (#29).
 * Examples in documentation of `p_attribute_encode()` have been adapted so that GitHub Actio unit test passes on Windows.
+* The `get_encoding()` method would return `NA` if `localeToCharset()` fails to infer charset from locale. In this case, UTF-8 is assumed.
 
 
 ## BUG FIXES